requirements.txt
tensorflow~=2.10.0
tqdm~=4.67.0
opencv-python~=4.10.0.84
numpy~=1.23.4
pillow~=11.0.0
scikit-learn~=1.2.2
matplotlib~=3.9.2
pillow~=11.0.0